v1.2.4
(Same as normal version without fullscreen)

-Fixed custom level uploading by increasing packet size to 250 and waiting 2 seconds every 15 packets sent so the send buffer doesn't overflow and also waiting 2 seconds before raising ReadyToLeave event
-Fixed Photon chat usernames not showing up
-Hide player usernames when silhouettes are disabled
-Added dropdown to main menu to pick cloud region to connect to
-Changed LevelObjectUpload event code since I think Photon is using event code 9 for something already
-Fixed back button in settings menus for team player prefabs
-Fixed white bots not going invisible by serializing and hard setting body, turret, barrel transforms in script instead of trying to find them at start

+++ Scripts/Runtime/Multiplayer/PhotonChatController.cs
@@ -647,7 +647,7 @@ public class PhotonChatController : MonoBehaviour, IChatClientListener
                 else
                 {
                     object[] messageData = (object[])messages[i];
-                    byte messageType = (byte)messageData[0];
+                    int messageType = (int)messageData[0];
                     if (messageType == 0)
                     {
                         IDUsernamePair.AddOrReplace((string)messageData[1], (string)messageData[2]);
